Relocate persistence code
