+++ cmd/pipeline/migrate.go
@@ -15,10 +15,11 @@
 package main
 
 import (
-	"github.com/banzaicloud/pipeline/internal/clusterfeature"
-	"github.com/banzaicloud/pipeline/internal/clustergroup/deployment"
 	"github.com/jinzhu/gorm"
 	"github.com/sirupsen/logrus"
+
+	"github.com/banzaicloud/pipeline/internal/clusterfeature/clusterfeatureadapter"
+	"github.com/banzaicloud/pipeline/internal/clustergroup/deployment"
 
 	"github.com/banzaicloud/pipeline/auth"
 	route53model "github.com/banzaicloud/pipeline/dns/route53/model"
@@ -83,7 +84,7 @@ func Migrate(db *gorm.DB, logger logrus.FieldLogger) error {
 		return err
 	}
 
-	if err := clusterfeature.Migrate(db, logger); err != nil {
+	if err := clusterfeatureadapter.Migrate(db, logger); err != nil {
 		return err
 	}
